refactor: remove OSSInsightClient implementation and related methods
- Deleted the entire OSSInsightClient class and its methods from the oss-insight-client.ts file.
- Updated actionItems.ts to call the GitHub API client directly for fetching good first issues and easy fixes without passing unnecessary parameters.
